[base] Move absl_status to base libraries
